refactor: put MLS functions in conversationService into their own files and just call them
