+++ packages/core/src/main/conversation/ConversationService/ConversationService.ts
@@ -17,8 +17,7 @@
  *
  */
 
-import type {CoreCrypto, Invitee} from '@otak/core-crypto';
-import {Decoder, Encoder} from 'bazinga64';
+import type {CoreCrypto} from '@otak/core-crypto';
 import type {APIClient} from '@wireapp/api-client';
 import {
   MessageSendingStatus,
@@ -95,13 +94,13 @@ import type {
 import {XOR} from '@wireapp/commons/src/main/util/TypeUtil';
 import type {NotificationService} from '../../notification';
 import {
-  AddUsersParams,
   MessageSendingCallbacks,
   MessageSendingOptions,
   MessageTargetMode,
-  SendMlsMessageParams,
   SendProteusMessageParams,
 } from './ConversationService.types';
+import {MLS} from './Protocol';
+import {AddUsersParams, SendMlsMessageParams} from './Protocol/MLS';
 
 export class ConversationService {
   public readonly messageTimer: MessageTimer;
@@ -109,11 +108,11 @@ export class ConversationService {
   private selfConversationId?: QualifiedId;
 
   constructor(
-    private readonly apiClient: APIClient,
+    protected readonly apiClient: APIClient,
     cryptographyService: CryptographyService,
     private readonly config: {useQualifiedIds?: boolean},
-    private readonly coreCryptoClientProvider: () => CoreCrypto,
-    private readonly notificationService: NotificationService,
+    protected readonly coreCryptoClientProvider: () => CoreCrypto,
+    protected readonly notificationService: NotificationService,
   ) {
     this.messageTimer = new MessageTimer();
     this.messageService = new MessageService(this.apiClient, cryptographyService);
@@ -877,98 +876,10 @@ export class ConversationService {
     }
 
     if (typeof conversationData !== 'string' && conversationData.protocol === ConversationProtocol.MLS) {
-      return this.createMLSConversation(conversationData);
+      return MLS.creatConversation.bind(this)(conversationData);
     }
 
     return this.apiClient.api.conversation.postConversation(payload);
-  }
-
-  private async createMLSConversation(conversationData: NewConversation): Promise<Conversation> {
-    /**
-     * @note For creating MLS conversations the users & qualified_users
-     * field must be empty as backend is not aware which users
-     * are in a MLS conversation because of the MLS architecture.
-     */
-    const newConversation = await this.apiClient.api.conversation.postConversation({
-      ...conversationData,
-      users: undefined,
-      qualified_users: undefined,
-    });
-    const {group_id: groupId} = newConversation;
-    const groupIdDecodedFromBase64 = Decoder.fromBase64(groupId!).asBytes;
-    const {qualified_users: qualifiedUsers = [], selfUserId} = conversationData;
-    if (!selfUserId) {
-      throw new Error('You need to pass self user qualified id in order to create an MLS conversation');
-    }
-    const coreCryptoClient = this.coreCryptoClientProvider();
-
-    await coreCryptoClient.createConversation(groupIdDecodedFromBase64);
-
-    const coreCryptoKeyPackagesPayload = await this.getCoreCryptoKeyPackagesPayload([
-      {
-        id: selfUserId.id,
-        domain: selfUserId.domain,
-        /**
-         * we should skip fetching key packages for current self client,
-         * it's already added by the backend on the group creation time
-         */
-        skipOwn: conversationData.creator_client,
-      },
-      ...qualifiedUsers,
-    ]);
-
-    await this.addUsersToExistingMLSConversation(groupIdDecodedFromBase64, coreCryptoKeyPackagesPayload);
-
-    await this.notificationService.saveConversationGroupId(newConversation);
-    // We fetch the fresh version of the conversation created on backend with the newly added users
-
-    return this.getConversations(newConversation.id);
-  }
-
-  private async getCoreCryptoKeyPackagesPayload(qualifiedUsers: (QualifiedId & {skipOwn?: string})[]) {
-    /**
-     * @note We need to fetch key packages for all the users
-     * we want to add to the new MLS conversations,
-     * includes self user too.
-     */
-    const keyPackages = await Promise.all([
-      ...qualifiedUsers.map(({id, domain, skipOwn}) =>
-        this.apiClient.api.client.claimMLSKeyPackages(id, domain, skipOwn),
-      ),
-    ]);
-
-    const coreCryptoKeyPackagesPayload = keyPackages.reduce<Invitee[]>((previousValue, currentValue) => {
-      // skip users that have not uploaded their MLS key packages
-      if (currentValue.key_packages.length > 0) {
-        return [
-          ...previousValue,
-          ...currentValue.key_packages.map(keyPackage => ({
-            id: Encoder.toBase64(keyPackage.client).asBytes,
-            kp: Decoder.fromBase64(keyPackage.key_package).asBytes,
-          })),
-        ];
-      }
-      return previousValue;
-    }, []);
-
-    return coreCryptoKeyPackagesPayload;
-  }
-
-  private async addUsersToExistingMLSConversation(groupIdDecodedFromBase64: Uint8Array, invitee: Invitee[]) {
-    const coreCryptoClient = this.coreCryptoClientProvider();
-    const memberAddedMessages = await coreCryptoClient.addClientsToConversation(groupIdDecodedFromBase64, invitee);
-    const sendingPromises: Promise<unknown>[] = [];
-    if (memberAddedMessages?.welcome) {
-      sendingPromises.push(
-        this.apiClient.api.conversation.postMlsWelcomeMessage(Uint8Array.from(memberAddedMessages.welcome)),
-      );
-    }
-    if (memberAddedMessages?.message) {
-      sendingPromises.push(
-        this.apiClient.api.conversation.postMlsMessage(Uint8Array.from(memberAddedMessages.message)),
-      );
-    }
-    await Promise.all(sendingPromises);
   }
 
   public async getConversations(conversationId: string): Promise<Conversation>;
@@ -1005,53 +916,15 @@ export class ConversationService {
     return response;
   }
 
-  private async addUsersToMLSGroup({conversationId, userIds, groupId}: AddUsersParams) {
-    const groupIdDecodedFromBase64 = Decoder.fromBase64(groupId!).asBytes;
-    console.info('addUsersToMLSGroup', conversationId, userIds, groupIdDecodedFromBase64);
-  }
-
   public async addUsers({conversationId, userIds, groupId}: AddUsersParams) {
     return groupId
-      ? this.addUsersToMLSGroup({conversationId, userIds, groupId})
+      ? MLS.addUsers.bind(this)({conversationId, userIds, groupId})
       : this.addUsersToProteusGroup({conversationId, userIds});
   }
 
   public async removeUser(conversationId: string, userId: string): Promise<string> {
     await this.apiClient.api.conversation.deleteMember(conversationId, userId);
     return userId;
-  }
-
-  private async sendMlsMessage<T extends OtrMessage>(
-    params: SendMlsMessageParams<T>,
-    genericMessage: GenericMessage,
-    content: T['content'],
-  ): Promise<T> {
-    const {groupId, onSuccess, payload} = params;
-    const groupIdBytes = Decoder.fromBase64(groupId).asBytes;
-
-    const coreCryptoClient = this.coreCryptoClientProvider();
-    const encrypted = await coreCryptoClient.encryptMessage(
-      groupIdBytes,
-      GenericMessage.encode(genericMessage).finish(),
-    );
-
-    try {
-      const {time = ''} = await this.apiClient.api.conversation.postMlsMessage(encrypted);
-      onSuccess?.(genericMessage, time?.length > 0 ? time : new Date().toISOString());
-      return {
-        ...payload,
-        content,
-        messageTimer: genericMessage.ephemeral?.expireAfterMillis || 0,
-        state: PayloadBundleState.OUTGOING_SENT,
-      };
-    } catch {
-      return {
-        ...payload,
-        content,
-        messageTimer: genericMessage.ephemeral?.expireAfterMillis || 0,
-        state: PayloadBundleState.CANCELLED,
-      };
-    }
   }
 
   private async sendProteusMessage<T extends OtrMessage>(
@@ -1110,7 +983,7 @@ export class ConversationService {
     }
 
     return isMLS(params)
-      ? this.sendMlsMessage(params, genericMessage, content)
+      ? MLS.sendMessage.bind(this)(params, genericMessage, content)
       : this.sendProteusMessage(params, genericMessage, content);
   }
